Move test buffer class into header so we can reuse it
