+++ tests/test-middle.cpp
@@ -12,7 +12,6 @@
 #include <algorithm>
 #include <cassert>
 
-#include <osmium/opl.hpp>
 #include <osmium/osm/crc.hpp>
 #include <osmium/osm/crc_zlib.hpp>
 
@@ -20,6 +19,7 @@
 #include "middle-pgsql.hpp"
 #include "middle-ram.hpp"
 
+#include "common-buffer.hpp"
 #include "common-cleanup.hpp"
 #include "common-options.hpp"
 #include "common-pg.hpp"
@@ -28,59 +28,12 @@ static testing::pg::tempdb_t db;
 
 namespace {
 
-/**
- * Wrapper around an Osmium buffer to create test objects in with some
- * convenience.
- */
-class test_buffer_t
-{
-public:
-    osmium::Node const &add_node(std::string const &data)
-    {
-        return m_buffer.get<osmium::Node>(add_opl(data));
-    }
-
-    osmium::Way &add_way(std::string const &data)
-    {
-        return m_buffer.get<osmium::Way>(add_opl(data));
-    }
-
-    osmium::Way &add_way(osmid_t wid, idlist_t const &ids)
-    {
-        assert(!ids.empty());
-        std::string nodes;
-
-        for (auto const id : ids) {
-            nodes += "n{},"_format(id);
-        }
-
-        nodes.resize(nodes.size() - 1);
-
-        return add_way("w{} N{}"_format(wid, nodes));
-    }
-
-    osmium::Relation const &add_relation(std::string const &data)
-    {
-        return m_buffer.get<osmium::Relation>(add_opl(data));
-    }
-
-private:
-    std::size_t add_opl(std::string const &data)
-    {
-        auto const offset = m_buffer.committed();
-        osmium::opl_parse(data.c_str(), m_buffer);
-        return offset;
-    }
-
-    osmium::memory::Buffer m_buffer{4096,
-                                    osmium::memory::Buffer::auto_grow::yes};
-};
-
 void expect_location(osmium::Location loc, osmium::Node const &expected)
 {
     CHECK(loc.lat() == Approx(expected.location().lat()));
     CHECK(loc.lon() == Approx(expected.location().lon()));
 }
+
 } // namespace
 
 struct options_slim_default
